+++ test/cli-utils-browser.spec.js
@@ -3,7 +3,7 @@
 
 const logger = require('winston');
 const mock = require('mock-require');
-const merge = require('../utils/merge-utils');
+const merge = require('../utils/merge');
 const url = require('url');
 
 const hostUtils = require('../utils/host-utils');
